Clean up core library python bindings
Distribute the monster python.cxx file out to the places where the relevant
structures are defined, maintaining the registration order of various
dependencies where necessary.
