+++ core/include/core/pybindings.h
@@ -97,209 +97,6 @@ register_vector_of(std::string name)
         return cls;
 }
 
-// Code to get a python buffer (for numpy) from one of our classes. "pyformat"
-// is the python type code for the type T.
-template <typename T>
-int pyvector_getbuffer(PyObject *obj, Py_buffer *view, int flags,
-    const char *pyformat)
-{
-	if (view == NULL) {
-		PyErr_SetString(PyExc_ValueError, "NULL view");
-		return -1;
-	}
-
-	view->shape = NULL;
-
-	boost::python::handle<> self(boost::python::borrowed(obj));
-	boost::python::object selfobj(self);
-	boost::python::extract<const std::vector<T> &> ext(selfobj);
-	if (!ext.check()) {
-		PyErr_SetString(PyExc_ValueError, "Invalid vector");
-		view->obj = NULL;
-		return -1;
-	}
-	const std::vector<T> &vec = ext();
-	view->obj = obj;
-	view->buf = (void*)&vec[0];
-	view->len = vec.size() * sizeof(T);
-	view->readonly = 0;
-	view->itemsize = sizeof(T);
-	if (flags & PyBUF_FORMAT)
-		view->format = (char *)pyformat;
-	else
-		view->format = NULL;
-	view->ndim = 1;
-#if (PY_MAJOR_VERSION == 2 && PY_MINOR_VERSION < 7) || (PY_MAJOR_VERSION >= 3)
-	// Abuse internal pointer in the absence of smalltable. This is safe
-	// on all architectures except MIPS N32.
-	view->internal = (void *)vec.size();
-	view->shape = (Py_ssize_t *)(&view->internal);
-#else
-	view->smalltable[0] = vec.size();
-	view->shape = &view->smalltable[0];
-	view->internal = NULL;
-#endif
-	view->strides = &view->itemsize;
-	view->suboffsets = NULL;
-
-	// Try to hold onto our collective hats. This is still very dangerous if
-	// the vector's underlying vector is resized.
-	Py_INCREF(obj);
-
-	return 0;
-}
-
-// For numeric vectors, instantiate this as container_from_object ahead of
-// registering the class.
-//
-// XXX: Should be automatic via SFINAE, but that cleanup is for later.
-template <typename T>
-boost::shared_ptr<T>
-numpy_container_from_object(boost::python::object v)
-{
-	// There's a chance this is actually a copy operation, so try that first
-	bp::extract<T &> extv(v);
-	if (extv.check())
-		return boost::make_shared<T>(extv());
-
-	boost::shared_ptr<T> x(new T);
-	size_t nelem;
-	Py_buffer view;
-
-	// See if we can get a numpy-ish buffer protocol interface next
-	if (PyObject_GetBuffer(v.ptr(), &view, PyBUF_FORMAT | PyBUF_STRIDES)
-	    == -1)
-		goto slowpython;
-
-	// Only 1D arrays here
-	if (view.ndim != 1)
-		goto bufferwasbad;
-
-	if (view.shape != NULL)
-		nelem = view.shape[0];
-	else
-		nelem = view.len/view.itemsize;
-	x->resize(nelem);
-
-	// Try to get a contiguous buffer first for common data types, since
-	// contiguous-copy is likely to be optimizable to memcpy(). In the
-	// non-contiguous case below, even if the copy is contiguous, the
-	// compiler won't see that. Note that Python's "contiguous" doesn't
-	// always mean exactly "contiguous", so we need to check strides too.
-	if (PyBuffer_IsContiguous(&view, 'A')) {
-		if (strcmp(view.format, "d") == 0) {
-			if (view.strides[0] != sizeof(double))
-				goto noncontiguous;
-			for (size_t i = 0; i < nelem; i++)
-				(*x)[i] = ((double *)view.buf)[i];
-			goto goodbuffer;
-		}
-	}
-
-noncontiguous:
-	if (strcmp(view.format, "d") == 0) {
-		for (size_t i = 0; i < nelem; i++)
-			(*x)[i] = *(double *)((char *)view.buf +
-			    i*view.strides[0]);
-	} else if (strcmp(view.format, "f") == 0) {
-		for (size_t i = 0; i < nelem; i++)
-			(*x)[i] = *(float *)((char *)view.buf +
-			    i*view.strides[0]);
-	} else if (strcmp(view.format, "n") == 0) {
-		for (size_t i = 0; i < nelem; i++)
-			(*x)[i] = *(ssize_t *)((char *)view.buf +
-			    i*view.strides[0]);
-	} else if (strcmp(view.format, "N") == 0) {
-		for (size_t i = 0; i < nelem; i++)
-			(*x)[i] = *(size_t *)((char *)view.buf +
-			    i*view.strides[0]);
-	} else if (strcmp(view.format, "?") == 0) { 
-		for (size_t i = 0; i < nelem; i++)
-			(*x)[i] = *(bool *)((char *)view.buf +
-			    i*view.strides[0]);
-	} else if (strcmp(view.format, "i") == 0) { 
-		for (size_t i = 0; i < nelem; i++)
-			(*x)[i] = *(int *)((char *)view.buf +
-			    i*view.strides[0]);
-	} else if (strcmp(view.format, "I") == 0) { 
-		for (size_t i = 0; i < nelem; i++)
-			(*x)[i] = *(unsigned int *)((char *)view.buf +
-			    i*view.strides[0]);
-	} else if (strcmp(view.format, "l") == 0) {
-		for (size_t i = 0; i < nelem; i++)
-			(*x)[i] = *(long *)((char *)view.buf +
-			    i*view.strides[0]);
-	} else if (strcmp(view.format, "L") == 0) {
-		for (size_t i = 0; i < nelem; i++)
-			(*x)[i] = *(unsigned long *)((char *)view.buf +
-			    i*view.strides[0]);
-	} else if (strcmp(view.format, "q") == 0) {
-		for (size_t i = 0; i < nelem; i++)
-			(*x)[i] = *(int64_t *)((char *)view.buf +
-			    i*view.strides[0]);
-	} else if (strcmp(view.format, "Q") == 0) {
-		for (size_t i = 0; i < nelem; i++)
-			(*x)[i] = *(uint64_t *)((char *)view.buf +
-			    i*view.strides[0]);
-	} else {
-		// We could add more types, but why do that?
-		// Let Python do the work for obscure cases
-		goto bufferwasbad;
-	}
-
-goodbuffer:
-	PyBuffer_Release(&view);
-        return x;
-
-bufferwasbad:
-	PyBuffer_Release(&view);
-
-slowpython:
-	// Failing all of that, try to do element-by-element conversion
-	// using Python iteration. If *that* fails, we just give up and
-	// go home (TypeError will be set inside extend_container()).
-	PyErr_Clear();
-	x->resize(0);
-	boost::python::container_utils::extend_container(*x, v);
-        
-        return x;
-}
-
-// Some special handling is needed for complex vectors
-
-template <typename T>
-boost::shared_ptr<T>
-complex_numpy_container_from_object(boost::python::object v)
-{
-	boost::shared_ptr<T> x(new T);
-	Py_buffer view;
-	if (PyObject_GetBuffer(v.ptr(), &view,
-	    PyBUF_FORMAT | PyBUF_ANY_CONTIGUOUS) != -1) {
-		if (strcmp(view.format, "Zd") == 0) {
-			x->resize(view.len/sizeof(std::complex<double>));
-			for (size_t i = 0; i < view.len/sizeof(std::complex<double>); i++)
-				(*x)[i] = ((std::complex<double> *)view.buf)[i];
-		} else if (strcmp(view.format, "Zf") == 0) {
-			x->resize(view.len/sizeof(std::complex<float>));
-			for (size_t i = 0; i < view.len/sizeof(std::complex<float>); i++)
-				(*x)[i] = ((std::complex<float> *)view.buf)[i];
-		} else {
-			// Fall back to scalar case otherwise
-			auto scalar =
-			   numpy_container_from_object<std::vector<double> >(v);
-			x->resize(scalar->size());
-			for (size_t i = 0; i < scalar->size(); i++)
-				(*x)[i] = (*scalar)[i];
-		}
-		PyBuffer_Release(&view);
-	} else {
-		PyErr_Clear();
-		boost::python::container_utils::extend_container(*x, v);
-	}
-
-	return x;
-}
-
 template <typename T, bool proxy=true>
 void
 register_map(std::string name, const char *docstring)
@@ -341,6 +138,8 @@ class G3ModuleRegistrator {
 public:
 	G3ModuleRegistrator(const char *mod, void (*def)());
 	static void CallRegistrarsFor(const char *mod);
+
+	SET_LOGGER("G3ModuleRegistrator");
 };
 
 #define EXPORT_G3MODULE_AND(mod, T, init, docstring, other_defs)   \
