+++ core/src/pybindings.cxx
@@ -1,0 +1,83 @@
+#include <pybindings.h>
+
+namespace bp = boost::python;
+
+// The following implements the headerless module registration code
+typedef std::map<std::string, std::vector<void (*)()> > module_reg_t;
+static module_reg_t *modregs = NULL;
+
+G3ModuleRegistrator::G3ModuleRegistrator(const char *mod, void (*def)())
+{
+	if (modregs == NULL)
+		modregs = new module_reg_t;
+	log_debug("Adding registrar for module %s", mod);
+	(*modregs)[mod].push_back(def);
+}
+
+void G3ModuleRegistrator::CallRegistrarsFor(const char *mod)
+{
+	for (auto i = (*modregs)[mod].begin(); i != (*modregs)[mod].end(); i++) {
+		log_debug("Calling registrar for module %s", mod);
+		(*i)();
+	}
+}
+
+G3PythonContext::G3PythonContext(std::string name, bool hold_gil) :
+    name_(name), hold_(false), thread_(nullptr)
+{
+	if (!Py_IsInitialized())
+		return;
+
+	if (hold_gil && !PyGILState_Check()) {
+		log_debug("%s: Ensuring GIL acquired", name_.c_str());
+		gil_ = PyGILState_Ensure();
+		hold_ = true;
+	} else if (!hold_gil && PyGILState_Check()) {
+		log_debug("%s: Saving Python thread state", name_.c_str());
+		thread_ = PyEval_SaveThread();
+	}
+}
+
+G3PythonContext::~G3PythonContext()
+{
+	if (hold_) {
+		log_debug("%s: Releasing GIL", name_.c_str());
+		PyGILState_Release(gil_);
+		hold_ = false;
+	}
+
+	if (!!thread_) {
+		log_debug("%s: Restoring Python thread state", name_.c_str());
+		PyEval_RestoreThread(thread_);
+		thread_ = nullptr;
+	}
+}
+
+G3PythonInterpreter::G3PythonInterpreter(bool hold_gil) :
+    init_(false)
+{
+	if (!Py_IsInitialized()) {
+		log_debug("Initializing");
+		Py_Initialize();
+#if (PY_MAJOR_VERSION == 3 && PY_MINOR_VERSION < 7)
+		PyEval_InitThreads();
+#endif
+		init_ = true;
+	}
+
+	ctx_ = new G3PythonContext("G3PythonInterpreter", hold_gil);
+}
+
+G3PythonInterpreter::~G3PythonInterpreter()
+{
+	if (!!ctx_) {
+		delete ctx_;
+		ctx_ = nullptr;
+	}
+
+	if (init_) {
+		log_debug("Finalizing");
+		Py_Finalize();
+		init_ = false;
+	}
+}
